+++ src/main/java/org/atatec/trugger/test/ScenarioFactory.java
@@ -1,18 +1,11 @@
 package org.atatec.trugger.test;
 
-import java.util.Collection;
-
 /**
- *
  * @author Marcelo Guimarães
  * @since 5.1
  */
 public interface ScenarioFactory {
 
-  ObjectScenario createObjectScenario(Object target);
-
-  IterableScenario createIterableScenario(Iterable target);
-
-  CollectionScenario createCollectionScenario(Collection target);
+  Scenario createObjectScenario(Object target);
 
 }